Fix padding of end of buffer character
The textarea component is incorrectly rendering the end of buffer
character when line numbers are enabled. Line number padding was still
being applied when displaying lines after the contents had been
rendered. This resulted in the following output:

```text
>   1 the first line
>   2 the second line
>   3 the third line
>   ~
>   ~
```

The correct functionality is based on how Vim renders this text which
should be as follows:

```text
>   1 the first line
>   2 the second line
>   3 the third line
> ~
> ~
```

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>